Fix #72 by avoiding 64-to-32-bit shortenings

+++ test/benchmark_test.cc
@@ -162,7 +162,7 @@ class TestReporter : public benchmark::internal::ConsoleReporter {
 
   virtual ~TestReporter() {}
 
-  int GetCount() const {
+  size_t GetCount() const {
     return count_;
   }
